fix: merge badge story data
