--- .storybook/mock-data/badge-tiered-data-mock.js
@@ -1,59 +1,0 @@
-export const badgeNoProgress = {
-	"id": "womens-equality",
-	"status": "PARTIALLY_COMPLETE",
-	"totalProgressToAchievement": 0,
-	"description": "description",
-	"matchingLoans": {
-		"filters": [
-			{
-				"gender": {
-					"eq": "female"
-				}
-			}
-		]
-	},
-	"tiers": [
-		{
-			"target": 20,
-			"learnMoreURL": "",
-			"completedDate": null,
-			"tierStatement": ""
-		},
-		{
-			"target": 50,
-			"learnMoreURL": "",
-			"completedDate": null,
-			"tierStatement": ""
-		},
-		{
-			"target": 100,
-			"learnMoreURL": "",
-			"completedDate": null,
-			"tierStatement": ""
-		},
-		{
-			"target": 2,
-			"learnMoreURL": "",
-			"completedDate": null,
-			"tierStatement": ""
-		},
-		{
-			"target": 3,
-			"learnMoreURL": "",
-			"completedDate": null,
-			"tierStatement": ""
-		},
-		{
-			"target": 5,
-			"learnMoreURL": "",
-			"completedDate": null,
-			"tierStatement": ""
-		},
-		{
-			"target": 10,
-			"learnMoreURL": "",
-			"completedDate": null,
-			"tierStatement": ""
-		}
-	]
-};
