+++ .storybook/mock-data/tiered-lending-achievement-data-mock.js
@@ -1,50 +1,110 @@
 export default [{
-  "id": "us-economic-equality",
-  "tiers": [
-    {
-      "completedDate": "2021-01-01T00:00:00Z",
-      "learnMoreURL": "",
-      "target": 2,
-      "tierStatement": ""
-    },
-    {
-      "completedDate": "2021-01-01T00:00:00Z",
-      "learnMoreURL": "",
-      "target": 3,
-      "tierStatement": ""
-    },
-    {
-      "completedDate": "2021-01-01T00:00:00Z",
-      "learnMoreURL": "",
-      "target": 5,
-      "tierStatement": ""
-    },
-    {
-      "completedDate": "2021-01-01T00:00:00Z",
-      "learnMoreURL": "",
-      "target": 10,
-      "tierStatement": ""
-    },
-    {
-      "completedDate": "2021-04-20T00:00:00Z",
-      "learnMoreURL": "https://www.kiva.org/impact/systemically-marginalized-communities",
-      "target": 20,
-      "tierStatement": "70% of U.S. business owners surveyed were more confident in their business as a result of their loan.*"
-    },
-    {
-      "completedDate": null,
-      "learnMoreURL": "",
-      "target": 50,
-      "tierStatement": ""
-    },
-    {
-      "completedDate": null,
-      "learnMoreURL": "",
-      "target": 100,
-      "tierStatement": ""
-    },
-  ],
-  "description": "description",
-  "totalProgressToAchievement": 0,
-  "status": "PARTIALLY_COMPLETE"
-}]
+	"id": "us-economic-equality",
+	"tiers": [
+		{
+			"completedDate": "2021-01-01T00:00:00Z",
+			"learnMoreURL": "",
+			"target": 2,
+			"tierStatement": ""
+		},
+		{
+			"completedDate": "2021-01-01T00:00:00Z",
+			"learnMoreURL": "",
+			"target": 3,
+			"tierStatement": ""
+		},
+		{
+			"completedDate": "2021-01-01T00:00:00Z",
+			"learnMoreURL": "",
+			"target": 5,
+			"tierStatement": ""
+		},
+		{
+			"completedDate": "2021-01-01T00:00:00Z",
+			"learnMoreURL": "",
+			"target": 10,
+			"tierStatement": ""
+		},
+		{
+			"completedDate": "2021-04-20T00:00:00Z",
+			"learnMoreURL": "https://www.kiva.org/impact/systemically-marginalized-communities",
+			"target": 20,
+			"tierStatement": "70% of U.S. business owners surveyed were more confident in their business as a result of their loan.*"
+		},
+		{
+			"completedDate": null,
+			"learnMoreURL": "",
+			"target": 50,
+			"tierStatement": ""
+		},
+		{
+			"completedDate": null,
+			"learnMoreURL": "",
+			"target": 100,
+			"tierStatement": ""
+		},
+	],
+	"description": "description",
+	"totalProgressToAchievement": 0,
+	"status": "PARTIALLY_COMPLETE"
+}];
+
+export const badgeNoProgress = {
+	"id": "womens-equality",
+	"status": "PARTIALLY_COMPLETE",
+	"totalProgressToAchievement": 0,
+	"description": "description",
+	"matchingLoans": {
+		"filters": [
+			{
+				"gender": {
+					"eq": "female"
+				}
+			}
+		]
+	},
+	"tiers": [
+		{
+			"target": 20,
+			"learnMoreURL": "",
+			"completedDate": null,
+			"tierStatement": ""
+		},
+		{
+			"target": 50,
+			"learnMoreURL": "",
+			"completedDate": null,
+			"tierStatement": ""
+		},
+		{
+			"target": 100,
+			"learnMoreURL": "",
+			"completedDate": null,
+			"tierStatement": ""
+		},
+		{
+			"target": 2,
+			"learnMoreURL": "",
+			"completedDate": null,
+			"tierStatement": ""
+		},
+		{
+			"target": 3,
+			"learnMoreURL": "",
+			"completedDate": null,
+			"tierStatement": ""
+		},
+		{
+			"target": 5,
+			"learnMoreURL": "",
+			"completedDate": null,
+			"tierStatement": ""
+		},
+		{
+			"target": 10,
+			"learnMoreURL": "",
+			"completedDate": null,
+			"tierStatement": ""
+		}
+	]
+};
